backend/requirements.txt
torch>=2.8.0
transformers>=4.57.0
datasets>=4.4.0
tokenizers>=0.22.0
accelerate>=1.11.0

soundfile>=0.13.0
torchaudio>=2.8.0

git+https://github.com/m-bain/whisperx.git
faster-whisper>=1.2.0
ctranslate2>=4.6.0
onnxruntime>=1.23.0

pyannote.audio>=3.4.0

python-dotenv>=1.0.0
sentencepiece>=0.2.0
tqdm>=4.67.0
numpy>=2.2.0
pandas>=2.2.0
scipy>=1.16.0
scikit-learn>=1.7.0

huggingface-hub>=0.36.0
safetensors>=0.6.0

openai>=0.27.0